arXiv-2303.10137/requirements.txt
torch
torchvision
rich_click
lightning
wandb